requirements.txt
catboost==1.2.7
pandas==2.2.3
protobuf==6.30.1
scikit_learn==1.6.1
streamlit==1.43.2
xgboost==1.2.1
lightgbm==2.3.1
google.cloud==0.34.0
google.oauth2==1.6.0

